+++ l2-frontend/src/forms/types/FileResultField.ts
@@ -1,4 +1,4 @@
-export type FileFieldValueFile = {
+export type ExistingFileFieldValueFile = {
   pk: number
   originalName: string
   extension: string
@@ -6,6 +6,19 @@ export type FileFieldValueFile = {
   size: number
   url: string
   createdAt: string
+  isNew?: false
 }
+
+export type NewFileFieldValueFile = {
+  tempId: string
+  originalName: string
+  extension: string
+  mimeType: string
+  size: number
+  file: File
+  isNew: true
+}
+
+export type FileFieldValueFile = ExistingFileFieldValueFile | NewFileFieldValueFile
 
 export type FileFieldValue = FileFieldValueFile[]
